homework-2/task3/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
pyfarmhash==0.4.0